+++ app/src/main/java/com/deskcontrol/DeskControlApp.kt
@@ -7,6 +7,8 @@ class DeskControlApp : Application() {
     override fun onCreate() {
         super.onCreate()
         SettingsStore.init(this)
+        DiagnosticsLog.init(resources)
+        SettingsStore.applyAppLanguage()
         AppCompatDelegate.setDefaultNightMode(SettingsStore.nightMode)
         DisplaySessionManager.init(this)
     }